RocksDB move deleteAllData() to test class
